requirements.txt
backports.functools-lru-cache==1.6.4
brotlipy==0.7.0
cached-property==1.5.2
email-validator==1.2.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-WTF==1.0.1
flit_core==3.7.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-resources==5.7.1
jupyter==1.0.0
jupyterlab==3.3.4
matplotlib==3.5.1
munkres==1.1.4
openpyxl==3.0.9
pandas==1.4.2
pip==22.0.4
pyOpenSSL==22.0.0
PySocks==1.7.1
scikit-learn==1.0.2
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
textblob==0.17.1
unicodedata2==14.0.0
